requirements.txt
langchain==0.1.8
python-decouple==3.8
langchain-openai==0.0.6